requirements-dev.txt
# Shekel Budget App -- Development and Testing Dependencies
#
# Includes all production dependencies plus tools for testing,
# linting, and local development.
#
# Install with:  pip install -r requirements-dev.txt

-r requirements.txt

# Testing
pytest==9.0.2
pytest-flask==1.3.0
pytest-cov==7.0.0
pytest-timeout==2.4.0
pytest-xdist==3.8.0
factory-boy==3.3.3

# Linting
pylint==4.0.5
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
